BaseTools: Using int instead of long
Using int instead of long

Cc: Liming Gao <liming.gao@intel.com>
Cc: Yonghong Zhu <yonghong.zhu@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1

+++ BaseTools/Source/Python/GenFds/FdfParser.py
@@ -1656,7 +1656,7 @@ class FdfParser:
                 self.SetPcdLocalation(pcdPair)
                 FileLineTuple = GetRealFileLine(self.FileName, self.CurrentLineNumber)
                 self.Profile.PcdFileLineDict[pcdPair] = FileLineTuple
-            Obj.Size = long(Size, 0)
+            Obj.Size = int(Size, 0)
             return True
 
         if self.__IsKeyword( "ErasePolarity"):
@@ -1692,7 +1692,7 @@ class FdfParser:
             if not self.__GetNextDecimalNumber() and not self.__GetNextHexNumber():
                 raise Warning("expected address", self.FileName, self.CurrentLineNumber)
 
-            BsAddress = long(self.__Token, 0)
+            BsAddress = int(self.__Token, 0)
             Obj.BsBaseAddress = BsAddress
 
         if self.__IsKeyword("RtBaseAddress"):
@@ -1702,7 +1702,7 @@ class FdfParser:
             if not self.__GetNextDecimalNumber() and not self.__GetNextHexNumber():
                 raise Warning("expected address", self.FileName, self.CurrentLineNumber)
 
-            RtAddress = long(self.__Token, 0)
+            RtAddress = int(self.__Token, 0)
             Obj.RtBaseAddress = RtAddress
 
     ## __GetBlockStatements() method
@@ -1750,7 +1750,7 @@ class FdfParser:
             self.SetPcdLocalation(PcdPair)
             FileLineTuple = GetRealFileLine(self.FileName, self.CurrentLineNumber)
             self.Profile.PcdFileLineDict[PcdPair] = FileLineTuple
-        BlockSize = long(BlockSize, 0)
+        BlockSize = int(BlockSize, 0)
 
         BlockNumber = None
         if self.__IsKeyword( "NumBlocks"):
@@ -1760,7 +1760,7 @@ class FdfParser:
             if not self.__GetNextDecimalNumber() and not self.__GetNextHexNumber():
                 raise Warning("expected block numbers", self.FileName, self.CurrentLineNumber)
 
-            BlockNumber = long(self.__Token, 0)
+            BlockNumber = int(self.__Token, 0)
 
         Obj.BlockSizeList.append((BlockSize, BlockNumber, BlockSizePcd))
         return True
@@ -1869,7 +1869,7 @@ class FdfParser:
             Expr += CurCh
             self.__GetOneChar()
         try:
-            return long(
+            return int(
                 ValueExpression(Expr,
                                 self.__CollectMacroPcd()
                                 )(True), 0)
@@ -1917,7 +1917,7 @@ class FdfParser:
                            RegionOffsetPcdPattern.match(self.__CurrentLine()[self.CurrentOffsetWithinLine:]))
             if IsRegionPcd:
                 RegionObj.PcdOffset = self.__GetNextPcdSettings()
-                self.Profile.PcdDict[RegionObj.PcdOffset] = "0x%08X" % (RegionObj.Offset + long(Fd.BaseAddress, 0))
+                self.Profile.PcdDict[RegionObj.PcdOffset] = "0x%08X" % (RegionObj.Offset + int(Fd.BaseAddress, 0))
                 self.SetPcdLocalation(RegionObj.PcdOffset)
                 self.__PcdDict['%s.%s' % (RegionObj.PcdOffset[1], RegionObj.PcdOffset[0])] = "0x%x" % RegionObj.Offset
                 FileLineTuple = GetRealFileLine(self.FileName, self.CurrentLineNumber)
@@ -3272,9 +3272,9 @@ class FdfParser:
                     if FdfParser.__Verify(Name, Value, 'UINT64'):
                         FmpData.MonotonicCount = Value
                         if FmpData.MonotonicCount.upper().startswith('0X'):
-                            FmpData.MonotonicCount = (long)(FmpData.MonotonicCount, 16)
+                            FmpData.MonotonicCount = (int)(FmpData.MonotonicCount, 16)
                         else:
-                            FmpData.MonotonicCount = (long)(FmpData.MonotonicCount)
+                            FmpData.MonotonicCount = (int)(FmpData.MonotonicCount)
             if not self.__GetNextToken():
                 break
         else:
